requirements.txt
Flask==3.0.0
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
websockets==12.0
aiohttp==3.9.1
joblib==1.3.2
prometheus-client==0.19.0
APScheduler==3.10.4
cryptography==41.0.7
pyjwt==2.8.0